# Coucher.WebApi/swagger-live.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Coucher.WebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/admin/closed-list-items": {
      "post": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClosedListItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClosedListItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClosedListItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/closed-list-items/bulk": {
      "post": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateClosedListItemRequest"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateClosedListItemRequest"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateClosedListItemRequest"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/closed-list-items/{id}": {
      "put": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClosedListItemRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClosedListItemRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClosedListItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/closed-list-items/display-orders": {
      "put": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateClosedListItemDisplayOrdersRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateClosedListItemDisplayOrdersRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/BulkUpdateClosedListItemDisplayOrdersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClosedListItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/closed-list-items/{id}/archive": {
      "post": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/closed-list-items/{id}/unarchive": {
      "post": {
        "tags": [
          "AdminClosedListItems"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "AUGUSTUS_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "MAMRAZ_CORRELATION_ID",
            "in": "header"
          },
          {
            "name": "AUGUSTUS_CORRELATION_CREATION_TIMESTAMP",
            "in": "header"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedListItem"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkUpdateClosedListItemDisplayOrdersRequest": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClosedListItemDisplayOrderUpdateRequest"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ClosedListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "key": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "value": {
            "maxLength": 256,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 1024,
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "isArchive": {
            "type": "boolean"
          },
          "creationTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastUpdateTime": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ClosedListItemDisplayOrderUpdateRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateClosedListItemRequest": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UpdateClosedListItemRequest": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "SessionId": {
        "type": "apiKey",
        "name": "session-id",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "SessionId": [ ]
    }
  ]
}